PMM-1507 Remove unused code.
